# api/openapi.yaml
openapi: 3.1.0
info:
  title: Labor Market Analytics API (LMA)
  description: |
    Публичный REST контракт BFF для React SPA (IT Labor Market Analytics).
    Это канонический контракт; docs/architecture/03-api.md описывает его семантику и внутреннее отображение.
    Salary-поля поддерживают RUB/USD/EUR/CNY/KZT/AMD; канон хранения остаётся RUB/net.
    AI- и Perspectives («Тенденции»)-маршруты помечены x-lifecycle: target.
  version: 1.0.0
  contact:
    name: LMA
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local BFF
  - url: https://{host}
    description: Deployed
    variables:
      host:
        default: api.example.com

# MVP read API is intentionally public. Admin operations override this with AdminToken.
security: []
...
paths:
...
  /api/v1/assistant/preferences:
    description: |
      In local/dev with ASSISTANT_DEV_AUTH_ENABLED=true, the BFF uses the
      deterministic server-side dev subject (or X-Dev-User when supplied).
      PUT, PATCH, and POST append an immutable preference version.
    get:
      tags: [assistant]
      operationId: getAssistantPreferences
      summary: Получить текущие персональные критерии
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Текущая версия критериев
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPreferences"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [assistant]
      operationId: saveAssistantPreferences
      summary: Сохранить новую версию критериев
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantPreferencesRequest"
      responses:
        "200":
          description: Сохранённая версия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPreferences"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/assistant/preferences/list:
    get:
      tags: [assistant]
      operationId: listAssistantPreferences
      summary: История версий критериев
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Версии, включая архивные
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssistantPreferences" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/assistant/preferences/archive:
    post:
      tags: [assistant]
      operationId: archiveAssistantPreference
      summary: Архивировать версию критериев
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string, format: uuid }
      responses:
        "204": { description: Версия архивирована; данные и evidence сохранены }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/assistant/status:
    get:
      tags: [assistant]
      operationId: getAssistantStatus
      summary: Статус последнего bounded анализа
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Durable status and aggregate counters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistantStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/assistant/analyze:
    post:
      tags: [assistant]
      operationId: runAssistantAnalysis
      summary: Поставить bounded анализ в очередь
      description: Не более 25 новых вакансий; Telegram не отправляется. Внешний AI требует отдельного server-side opt-in.
      security:
        - bearerAuth: []
      responses:
        "202": { description: Queued }
        "409": { description: Уже выполняется }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/assistant/matches:
    get:
      tags: [assistant]
      operationId: listAssistantMatches
      summary: Получить персональные совпадения
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Совпадения
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssistantMatch"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/assistant/telegram/link:
    post:
      tags: [assistant]
      operationId: createTelegramLink
      summary: Создать одноразовую ссылку Telegram
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deep-link без bot token
          content:
            application/json:
              schema:
                type: object
                required: [deep_link, expires_at]
                properties:
                  deep_link: { type: string, format: uri }
                  expires_at: { type: string, format: date-time }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/assistant/telegram:
    get:
      tags: [assistant]
      operationId: getTelegramStatus
      summary: Статус Telegram connection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Статус
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [assistant]
      operationId: revokeTelegram
      summary: Отозвать Telegram connection
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Отозвано
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [assistant]
      operationId: updateTelegramOptIn
      summary: Изменить отдельное согласие на Telegram-уведомления
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opted_in]
              properties:
                opted_in: { type: boolean }
      responses:
        "200":
          description: Согласие изменено
          content:
            application/json:
              schema:
                type: object
                required: [opted_in]
                properties:
                  opted_in: { type: boolean }
        "409":
          description: Telegram не связан и не подтверждён
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/assistant/telegram/test:
    post:
      tags: [assistant]
      operationId: sendTelegramTestNotification
      summary: Отправить явно подтверждённое тестовое уведомление
      description: Реальная отправка возможна только после verified link и opt-in; chat_id из клиента не принимается.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm: { type: boolean, const: true }
      responses:
        "200": { description: Отправлено }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Telegram не подтверждён }
        "503": { description: Telegram sender не настроен }
  /api/v1/assistant/automation:
    get:
      tags: [assistant]
      operationId: getAssistantAutomation
      summary: Получить настройки автоматического анализа
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Персональные opt-in настройки и activation cutoff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistantAutomationSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [assistant]
      operationId: updateAssistantAutomation
      summary: Включить или выключить AI и Telegram отдельно
      description: Включение AI фиксирует activation_at; старые вакансии не backfill-ятся.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ai_enabled: { type: boolean }
                telegram_enabled: { type: boolean }
                max_ai_calls_per_hour: { type: integer, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: Сохранённые настройки
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistantAutomationSettings" }
        "409":
          description: Серверный provider не настроен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: MVP admin gate; Target — Bearer JWT
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              examples: [VALIDATION_ERROR, CONFLICT, RATE_LIMITED, DEPENDENCY_UNAVAILABLE]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
...
    AssistantPreferencesRequest:
      type: object
      required: [hard_criteria, soft_criteria, weights]
      properties:
        note: { type: string, maxLength: 2000 }
        hard_criteria:
          type: object
          additionalProperties: true
        soft_criteria:
          type: object
          additionalProperties: true
        weights:
          type: object
          additionalProperties:
            type: number
            minimum: 0
    AssistantPreferences:
      allOf:
        - $ref: "#/components/schemas/AssistantPreferencesRequest"
        - type: object
          required: [version, active_from]
          properties:
            id: { type: string, format: uuid }
            version: { type: integer, minimum: 1 }
            active_from: { type: string, format: date-time }
            archived_at: { type: [string, "null"], format: date-time }
    AssistantStatus:
      type: object
      required: [ai_configured, state, processed, eligible, matched, ai_calls, skipped, pending_candidates]
      properties:
        ai_configured: { type: boolean }
        state: { type: string, enum: [never_run, queued, running, succeeded, failed, disabled] }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        last_checked_at: { type: [string, "null"], format: date-time }
        processed: { type: integer, minimum: 0 }
        eligible: { type: integer, minimum: 0 }
        matched: { type: integer, minimum: 0 }
        ai_calls: { type: integer, minimum: 0 }
        skipped: { type: integer, minimum: 0 }
        error_category: { type: [string, "null"] }
        request_id: { type: [string, "null"] }
        cursor_source: { type: string }
        cursor_observed_at: { type: [string, "null"], format: date-time }
        pending_candidates: { type: boolean }
        provider: { type: [string, "null"] }
        model: { type: [string, "null"] }
        prompt_version: { type: [string, "null"] }
        method_version: { type: string }
    AssistantMatch:
      type: object
      required: [vacancy_id, decision, score, method, reasons, unknowns, conflicts, evidence_ids]
      properties:
        vacancy_id: { type: string, format: uuid }
        title: { type: string }
        source_url: { type: [string, "null"], format: uri }
        decision: { type: string, enum: [match, reject, review] }
        score: { type: number, minimum: 0, maximum: 1 }
        method: { type: string, enum: [deterministic, ai] }
        confidence: { type: [string, "null"], enum: [low, medium, high, null] }
        reasons: { type: array, items: { type: string } }
        unknowns: { type: array, items: { type: string } }
        conflicts: { type: array, items: { type: string } }
        evidence_ids: { type: array, items: { type: string } }
    TelegramStatus:
      type: object
      required: [configured, linked, opted_in, pending, sent, failed, dead_lettered, last_error]
      properties:
        configured: { type: boolean }
        linked: { type: boolean }
        opted_in: { type: boolean }
        pending: { type: integer, minimum: 0 }
        sent: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
        dead_lettered: { type: integer, minimum: 0 }
        last_error: { type: string }
    AssistantAutomationSettings:
      type: object
      required: [ai_enabled, telegram_enabled, max_ai_calls_per_hour]
      properties:
        ai_enabled: { type: boolean }
        telegram_enabled: { type: boolean }
        activation_at:
          type: [string, "null"]
          format: date-time
          description: Cutoff включения; вакансии, наблюдённые ранее, не backfill-ятся.
        max_ai_calls_per_hour:
          type: integer
          minimum: 1
          maximum: 100